### 2017-03-07 1196587
DOC : Typos in advanced hdbscan
