### 2017-03-07 b3d07e4
DOC : Cluster trees for condensed trees.
