### 2017-03-07 679a5fa
DOC : Single linkage tree get_clusters.
