+++ setup.py
@@ -37,7 +37,7 @@ def readme():
 
 configuration = {
     'name' : 'hdbscan',
-    'version' : '0.8.7',
+    'version' : '0.8.8',
     'description' : 'Clustering based on density with variable density clusters',
     'long_description' : readme(),
     'classifiers' : [
