### 2017-03-07 d65eca1
DOC : Add extra modules to API docs.
