### 2017-03-07 15be9c4
Add exemplars as a property of basic HDBSCAN objects.
