### 2017-03-07 828317f
DOC : exemplars property doctsring.
