### 2017-03-08 584c1c0
DOC : Tidy up plots docstring.
