### 2017-03-08 4ee3a19
DOC : Add joblib link in parameter selection.
