### 2017-03-08 90b6885
DOC : Typos and truth in parameter selection docs.
