### 2017-03-08 15f3a6e
DOC : Autoclass the plots classes
