### 2017-03-08 b1ef1e5
DOC : Formatting in outlier detection.
