### 2017-03-08 134004b
DOC : Leaf clustering description.
