### 2017-03-08 b345c1c
DOC : Allowing single clusters.
