+++ docs/parameter_selection.rst
@@ -8,6 +8,8 @@ parameters that have significant practical effect on clustering. We will
 consider those major parameters, and consider how one may go about
 choosing them effectively.
 
+.. _min_cluster_size_label
+
 Selecting ``min_cluster_size``
 ------------------------------
 
@@ -117,6 +119,8 @@ computation, and recompute only the relatively cheap flat cluster
 extraction using the ``memory`` parameter, which makes use of
 `joblib <https://pythonhosted.org/joblib/>`_
 
+.. _min_samples_label
+
 Selecting ``min_samples``
 -----------------------
 
@@ -160,6 +164,8 @@ above, make the clustering progressively more conservative, culminating
 in the example above where ``min_samples`` was set to 60 and we had only
 two clusters with most points declared as noise.
 
+.. _alpha_label
+
 Selecting ``alpha``
 -----------------
 
@@ -188,6 +194,9 @@ as we can see by setting ``alpha`` to 1.3.
 
 .. image:: images/parameter_selection_18_1.png
 
+
+.. _leaf_clustering_label
+
 Leaf clustering
 ---------------
 
@@ -205,6 +214,8 @@ homogeneous clusters. Note that you can still get variable density clusters
 via this method, and it is also still possible to get large clusters, but
 there will be a tendency to produce a more fine grained clustering than
 Excess of Mass can provide.
+
+.. _single_cluster_label
 
 Allowing a single cluster
 -------------------------
