### 2017-03-08 cb2c63d
DOC : Generative model comparison in FAQ.
