### 2017-03-08 a79bf8c
DOC : Predict in the FAQ.
