### 2017-03-08 20c6ed0
DOC : Minor FAQ cleanup.
