### 2017-03-09 8827b6d
DOC : Remove warning output
