### 2017-03-09 1bee94d
DOC : reference approximate predict function
