### 2017-03-09 2de91d4
Fix spelling errors in how_hdbscan_works.rst
