### 2017-03-09 4dcd9c3
DOC : clean up spurious backticks.
