### 2017-03-10 93e4726
DOC : CondensedTree class docstring
