### 2017-03-10 bf41308
DOC : SingleLinkageTree class docstring.
