### 2017-03-10 d4c2a99
DOC : get_clusters docstring.
