### 2017-03-10 37d9b8d
DOC : Return values for get_clusters in docstring.
