### 2017-03-10 f5b82c4
DOC : outlier_scores docstring
