### 2017-03-10 8225139
DOC : labelling_at_cut docstring.
