### 2017-03-10 91344d3
DOC : condense_tree docstring.
