### 2017-03-10 7befb0d
DOC : FAQ on haversine.
