### 2017-03-10 f5e8157
No need for weighting between dist and outlier vecs.
