### 2017-03-10 0d30b88
Go back to basic GLOSH style relative outlier score.
