### 2017-03-10 583c674
DOC : Add How soft clustering works to the docs
