+++ docs/soft_clustering_explanation.rst
@@ -1,0 +1,493 @@
+
+How Soft Clustering for HDBSCAN Works
+=====================================
+
+This is a general description of how the soft clustering algorithm for
+HDBSCAN Works. We will implement soft clustering from scratch -- not in
+the efficient way that the `hdbscan
+library <https://github.com/scikit-learn-contrib/hdbscan>`__ implements
+it, but in a way that makes it clearer what is actually going on.
+
+What is Soft Clustering?
+------------------------
+
+To start, we'll provide a quick primer on what soft clustering is, and
+why you might want it in the first place. Traditional clustering assigns
+each point in a data set to a cluster (or to noise). This is a hard
+assignment; there are no mixed memberships. A point near the edge of one
+cluster and also close to a second cluster, is just as much "in the
+first cluster" as a point solidly in the center that is very distant
+from the second cluster. Equally, if the clustering algorithm supports
+noise assignments, then points are simply assigned as "noise". We are
+left with no idea as to which, if any cluster, they might have just
+missed the cur on being in.
+
+The remedy for this is 'soft clustering' or 'fuzzy clustering'. In this
+approach points are no assigned cluster labels, but are instead assigned
+a vector of probabilities. The length of the vector is equal to the
+number of clusters found. The probability value at the :math:`i`\ th
+entry of the vector is the probability that that point is a member of
+the :math:`i`\ th cluster. This allows points to potentially be a mix of
+clusters. By looking at the vector a data scientist can discern how
+strongly a point is in a cluster, and which other clusters it is related
+to. Equally noise point will usually be assigned low probabilities of
+being in any cluster, but you can discern which clusters they are closer
+to, or even if they were very nearly part of a cluster. This is a much
+richer and more informative clustering result ... so how do we get it?
+
+Soft Clustering for HDBSCAN
+---------------------------
+
+We want to produce a method of providing a soft membership vector for a
+given point across the selected clusters of a clustering. Ideally we
+want this to be interpretable as a probability of being a member of that
+cluster. For now we will work solely with categorizing points already in
+the clustered data set, but in principle this can be extended to new
+previously unseen points presuming we have a method to insert such
+points into the condensed tree (see other discussions on how to handle
+"prediction"). First let's get some test data, and an associated
+clustering so we have some idea of what we're working with in our test
+example.
+
+.. code:: python
+
+    import hdbscan
+    import numpy as np
+    import matplotlib.pyplot as plt
+    import seaborn as sns
+    import pandas as pd
+    import matplotlib as mpl
+    
+    from scipy.spatial.distance import cdist
+    
+    %matplotlib inline
+    sns.set_context('poster')
+    sns.set_style('white')
+    sns.set_color_codes()
+    
+    plot_kwds={'alpha':0.25, 's':60, 'linewidths':0}
+    palette = sns.color_palette('deep', 12)
+
+
+.. code:: python
+
+    data = np.load('clusterable_data.npy')
+    fig = plt.figure()
+    ax = fig.add_subplot(111)
+    plt.scatter(data.T[0], data.T[1], **plot_kwds)
+    ax.set_xticks([])
+    ax.set_yticks([]);
+
+
+
+.. image:: How%20Soft%20Clustering%20for%20HDBSCAN%20Works_files/How%20Soft%20Clustering%20for%20HDBSCAN%20Works_2_0.png
+
+
+Now let's build a clusterer and fit it to this data.
+
+.. code:: python
+
+    clusterer = hdbscan.HDBSCAN(min_cluster_size=15).fit(data)
+
+We can visualize the resulting clustering (using the soft cluster scores
+to vary the saturation so that we gain some intuition about how soft the
+clusters may be) to get an idea of what we are looking at:
+
+.. code:: python
+
+    pal = sns.color_palette('deep', 8)
[…]
