### 2017-03-10 9e826f9
Membership vector tests need to be updated.
