### 2017-03-10 b7ea79f
DOC : soft clustering doc image paths
