### 2017-03-10 acd69df
DOC : soft clustering doc typos
