### 2017-03-10 3aba57f
DOC : Add soft clustering explanation to index
