### 2017-03-10 cf3cffd
Fix up prediction utils.
