+++ hdbscan/_prediction_utils.c
@@ -936,7 +936,7 @@ struct __pyx_opt_args_7hdbscan_17_prediction_utils_all_points_dist_membership_ve
   PyObject *softmax;
 };
 
-/* "hdbscan/_prediction_utils.pyx":221
+/* "hdbscan/_prediction_utils.pyx":219
  *     return result
  * 
  * cpdef np.ndarray[np.float64_t, ndim=1] outlier_membership_vector(neighbor,             # <<<<<<<<<<<<<<
@@ -948,7 +948,7 @@ struct __pyx_opt_args_7hdbscan_17_prediction_utils_outlier_membership_vector {
   PyObject *softmax;
 };
 
-/* "hdbscan/_prediction_utils.pyx":300
+/* "hdbscan/_prediction_utils.pyx":298
  *     return result_arr
  * 
  * cpdef np.ndarray[np.float64_t, ndim=2] all_points_outlier_membership_vector(             # <<<<<<<<<<<<<<
@@ -4758,38 +4758,38 @@ static PyArrayObject *__pyx_f_7hdbscan_17_prediction_utils_per_cluster_scores(__
   __pyx_v_result = ((PyArrayObject *)__pyx_t_1);
   __pyx_t_1 = 0;
 
-  /* "hdbscan/_prediction_utils.pyx":216
+  /* "hdbscan/_prediction_utils.pyx":213
  * 
  *     # Cythonize: result = np.exp(-(max_lambda / height))
  *     for i in range(result.shape[0]):             # <<<<<<<<<<<<<<
- *         result[i] = exp(-(max_lambda / result[i]))
- * 
+ *         # result[i] = exp(-(max_lambda / result[i]))
+ *         result[i] = (max_lambda / (max_lambda - result[i]))
  */
   __pyx_t_9 = (__pyx_v_result->dimensions[0]);
   for (__pyx_t_2 = 0; __pyx_t_2 < __pyx_t_9; __pyx_t_2+=1) {
     __pyx_v_i = __pyx_t_2;
 
-    /* "hdbscan/_prediction_utils.pyx":217
- *     # Cythonize: result = np.exp(-(max_lambda / height))
+    /* "hdbscan/_prediction_utils.pyx":215
  *     for i in range(result.shape[0]):
- *         result[i] = exp(-(max_lambda / result[i]))             # <<<<<<<<<<<<<<
+ *         # result[i] = exp(-(max_lambda / result[i]))
+ *         result[i] = (max_lambda / (max_lambda - result[i]))             # <<<<<<<<<<<<<<
  * 
  *     return result
  */
     __pyx_t_10 = __pyx_v_i;
     if (__pyx_t_10 < 0) __pyx_t_10 += __pyx_pybuffernd_result.diminfo[0].shape;
-    __pyx_t_4 = (*__Pyx_BufPtrStrided1d(__pyx_t_5numpy_float64_t *, __pyx_pybuffernd_result.rcbuffer->pybuffer.buf, __pyx_t_10, __pyx_pybuffernd_result.diminfo[0].strides));
+    __pyx_t_4 = (__pyx_v_max_lambda - (*__Pyx_BufPtrStrided1d(__pyx_t_5numpy_float64_t *, __pyx_pybuffernd_result.rcbuffer->pybuffer.buf, __pyx_t_10, __pyx_pybuffernd_result.diminfo[0].strides)));
[…]
