### 2017-03-10 6b25d47
DOC : Add soft clustering doc
