+++ docs/soft_clustering.rst
@@ -1,0 +1,217 @@
+
+Soft Clustering for HDBSCAN\*
+=============================
+
+Soft clustering is a new (and still somewhat experimental) feature of
+the hdbscan library. It takes advantage of the fact that the condensed
+tree is a kind of smoothed density function over data points, and the
+notion of exemplars for clusters. If you want to better understand how
+soft clustering works please refer to How Soft Clustering for HDBSCAN\*
+Works.
+
+Let's consider the digits dataset from sklearn. We can project the data
+into two dimensions to visualize it via t-SNE.
+
+.. code:: python
+
+    from sklearn import datasets
+    from sklearn.manifold import TSNE
+    import matplotlib.pyplot as plt
+    import seaborn as sns
+    import numpy as np
+    
+    %matplotlib inline
+    
+    sns.set_context('poster')
+    sns.set_style('white')
+    sns.set_color_codes()
+    
+    plot_kwds={'alpha':0.25, 's':60, 'linewidths':0}
+
+.. code:: python
+
+    digits = datasets.load_digits()
+    data = digits.data
+    projection = TSNE().fit_transform(data)
+    plt.scatter(*projection.T, **plot_kwds)
+
+
+
+
+.. parsed-literal::
+
+    <matplotlib.collections.PathCollection at 0x112d39978>
+
+
+
+
+.. image:: soft_clustering_files/soft_clustering_3_1.png
+
+
+Now we import hdbscan and then cluster in the full 64 dimensional space.
+It is important to note that, if we wish to use the soft clustering we
+should use the ``prediction_data=True`` option for HDBSCAN. This will
+ensure we generate the extra data required that will allow soft
+clustering to work.
+
+.. code:: python
+
+    import hdbscan
+
+.. code:: python
+
+    clusterer = hdbscan.HDBSCAN(min_cluster_size=10, prediction_data=True).fit(data)
+    color_palette = sns.color_palette('Paired', 12)
+    cluster_colors = [color_palette[x] if x >= 0 
+                      else (0.5, 0.5, 0.5) 
+                      for x in clusterer.labels_]
+    cluster_member_colors = [sns.desaturate(x, p) for x, p in 
+                             zip(cluster_colors, clusterer.probabilities_)]
+    plt.scatter(*projection.T, s=50, linewidth=0, c=cluster_member_colors, alpha=0.25)
+
+
+
+
+.. parsed-literal::
+
+    <matplotlib.collections.PathCollection at 0x112a80e10>
+
+
+
+
+.. image:: soft_clustering_files/soft_clustering_6_1.png
+
+
+Certainly a number of clusters were found, but the data is fairly noisy
+in 64 dimensions, so there are a number of points that have been
+classified as noise. We can generate a soft clustering to get more
+information about some of these noise points.
+
+To generate a soft clustering for all the points in the original dataset
+we use the
+:py:func:``~hdbscan.prediction.all_points_membership_vectors`` function
+which takes a clusterer object. If we wanted to get soft cluster
+membership values for a set of new unseen points we could use
+:py:func:``~hdbscan.prediction.membership_vector`` instead.
+
+The return value is a two-dimensional numpy array. Each point of the
+input data is assigned a vector of probabilities of being in a cluster.
+For a first pass we can visualize the data looking at what the *most
+likely* cluster was, by coloring according to the ``argmax`` of the
+probability vector (i.e. the cluster for which a given point has the
+highest probability of being in).
+
+.. code:: python
+
+    soft_clusters = hdbscan.all_points_membership_vectors(clusterer)
+    color_palette = sns.color_palette('Paired', 12)
+    cluster_colors = [color_palette[np.argmax(x)]
[…]
