### 2017-03-10 a600376
DOC : Clean up soft clustering.
