### 2017-03-10 8f3a879
DOC : Add osft clustering to index.
