### 2017-03-10 8aabf7b
DOC : Image paths in soft clustering doc
