+++ docs/prediction_tutorial.rst
@@ -1,0 +1,151 @@
+
+Predicting clusters for new points
+==================================
+
+Often it is useful to train a model once on a large amount of data, and
+then query the model repeatedly with small amounts of new data. This is
+hard for HDBSCAN\* as it is a transdusctive method -- new data points
+can (and should!) be able to alter the underlying clustering. That is,
+given new information it might make sense to create a new cluster, split
+an existing cluster, or merge two previously separate clusters. If the
+actual clusters (and hence their labels) change with each new data point
+it becomes impossible to compare the cluster assignments between such
+queries.
+
+We can accomodate this by effectively holding a clustering fixed (after
+a potentially expensive training run) and then asking: *if we do not
+change the existing clusters* which cluster would HDBSCAN\* assign a new
+data point to. In practice this amounts to determining where in the
+condensed tree the new data point would fall (see
+:any:``how_hdbscan_works``) assuming we do not change the condensed
+tree. This allows for a very inexpensive operation to compute a
+predicted cluster for the new data point.
+
+This has been implemented in ``hdbscan`` as the
+:py:func:``~hdbscan.predict.approximate_predict`` function. We'll look
+at how this works below.
+
+.. code:: python
+
+    import hdbscan
+    import numpy as np
+    import matplotlib.pyplot as plt
+    import seaborn as sns
+    import matplotlib as mpl
+    from sklearn.neighbors import KDTree, BallTree
+    from sklearn.datasets import make_blobs
+    
+    %matplotlib inline
+    sns.set_context('poster')
+    sns.set_style('white')
+    sns.set_color_codes()
+    
+    plot_kwds={'alpha':0.25, 's':60, 'linewidths':0}
+    palette = sns.color_palette('deep', 12)
+
+As usual we begin with our test synthetic data set, and cluster it with
+HDBSCAN. The primary point to note here, however, is the use of the
+``prediction_data=True`` keyword argument. This ensures that HDBSCAN
+does a little extra computation when fitting the model that can
+dramatically speed up the prediction queries later.
+
+You can also get an HDBSCAN object to create this data after the fact
+via the :py:meth:``~hdbscan.HDBSCAN.generate_prediction_data`` method.
+
+.. code:: python
+
+    data = np.load('clusterable_data.npy')
+    clusterer = hdbscan.HDBSCAN(min_cluster_size=15, prediction_data=True).fit(data)
+    pal = sns.color_palette('deep', 8)
+    colors = [sns.desaturate(pal[col], sat) for col, sat in zip(clusterer.labels_, 
+                                                                clusterer.probabilities_)]
+    plt.scatter(data.T[0], data.T[1], c=colors, **plot_kwds);
+
+
+
+.. image:: prediction_tutorial_files/prediction_tutorial_3_0.png
+
+
+Now to make things a little more interesting let's generate 50 new data
+points scattered across the data. We can plot them in black to see where
+they happen to fall.
+
+.. code:: python
+
+    test_points = np.random.random(size=(50, 2)) - 0.5
+    
+    colors = [sns.desaturate(pal[col], sat) for col, sat in zip(clusterer.labels_, 
+                                                                clusterer.probabilities_)]
+    plt.scatter(data.T[0], data.T[1], c=colors, **plot_kwds);
+    plt.scatter(*test_points.T, c='k', s=50)
+
+
+
+
+.. parsed-literal::
+
+    <matplotlib.collections.PathCollection at 0x115af1908>
+
+
+
+
+.. image:: prediction_tutorial_files/prediction_tutorial_5_1.png
+
+
[…]
