### 2017-03-10 7340940
DOC : Clean up prediction tutorial
