### 2017-03-10 8694921
DOC : image paths in prediction tutorial
