### 2017-03-10 4d5e69e
DOC : soft clustering: clean up backticks.
