### 2017-03-10 6d1dae3
DOC : Clean up links in prediction tutorial.
