### 2017-03-10 0b56b0b
DOC : Add prediction tutorial to index
