### 2017-03-10 70e01b8
DOC : Minor docstring fix -- add references.
