### 2017-03-13 #191 @danielskatz
Hi @lmcinnes
While I look for a reviewer, you might want to fix the formatting of the paper part of your submission.  See https://github.com/mpastell/Weave.jl/blob/master/paper/paper.md for an example of another current submission that has a properly formatted paper.  By this, I refer to two things:
1. the metadata, which looks ok to me, but doesn't seem to generate a proper table like I expected, though perhaps this is a parsing issue, since I don't see anything wrong with your metadata itself.  Ah, I see it now - your line 12 needs one more space at the start.
2. references to items in the bibliography.  Only items that are referenced in the text will be shown in the bibliography when the paper is compiled, and you currently don't refer to any of them.

Thanks.
