### 2017-03-13 #205 @whedon
[REVIEW]: hdbscan: A high performance implementation of HDBSCAN* clustering. 
**Submitting author:** @lmcinnes (<a href="http://orcid.org/0000-0003-2143-6834">Leland McInnes</a>)
**Repository:** <a href="https://github.com/scikit-learn-contrib/hdbscan" target ="_blank">https://github.com/scikit-learn-contrib/hdbscan</a>
**Version:** v0.8.8
