### 2017-03-13 cde9bfe
Fix formatting of metadata table
