### 2017-03-13 #205 @lmcinnes
Hi @dskatz

Thanks for taking on the review! I'm honoured to have such esteemed
reviewers for this.

I have fixed the metadata as requested. Sorry for the oversight. I believe
I have provided references, hopefully in the correct format (based off the
example).

Thanks,

Leland McInnes.

On Mon, Mar 13, 2017 at 10:04 AM, Daniel S. Katz <notifications@github.com>
wrote:

> Hi @lmcinnes <https://github.com/lmcinnes>
>
> Please fix the formatting of the paper part of your submission. See
> https://github.com/mpastell/Weave.jl/blob/master/paper/paper.md for an
> example of another current submission that has a properly formatted paper.
> By this, I refer to two things:
>
>    1. the metadata: your line 12 needs one more space at the start.
>    2. references to items in the bibliography. Only items that are
>    referenced in the text will be shown in the bibliography when the paper is
>    compiled, and you currently don't refer to any of them.
>
> Thanks.
>
