### 2017-03-13 #205 @danielskatz
Thanks - there were also a couple more things I saw, so I put in a pull request.
