### 2017-03-17 #205 @zhaozhang
Hi @lmcinnes,

I have found the following points during installation:
  1. hdbscan depends on scipy also, it was not documented in the instruction manual.
  2. it works with both Python2 and Python3, please document this.
  3. The code examples in README.md are not self-contained, please replace "data" with some example dataset.

Please also include community guidelines:
  1. Contribute to the software.
  2. Report issues or problems with the software.
  3. Seek support.

Please add DOI for the references in paper.bib.

Best,
Zhao
