### 2017-03-18 3945b3e
Update README as per suggestions from @zhaozhang
