+++ README.rst
@@ -60,6 +60,9 @@ giving a distance matrix between samples.
 .. code:: python
 
     import hdbscan
+    from sklearn.datasets import make_blobs
+    
+    data = make_blobs(1000)
     
     clusterer = hdbscan.HDBSCAN(min_cluster_size=10)
     cluster_labels = clusterer.fit_predict(data)
@@ -131,6 +134,9 @@ Example usage:
 .. code:: python
 
     import hdbscan
+    from sklearn.datasets import make_blobs
+    
+    data = make_blobs(1000)
     
     clusterer = hdbscan.RobustSingleLinkage(cut=0.125, k=7)
     cluster_labels = clusterer.fit_predict(data)
@@ -154,7 +160,7 @@ Easiest install, if you have Anaconda (thanks to conda-forge which is awesome!):
 
     conda install -c conda-forge hdbscan
 
-PyPI install, presuming you have sklearn and all its requirements installed:
+PyPI install, presuming you have sklearn and all its requirements (numpy and scipy) installed:
 
 .. code:: bash
 
@@ -166,6 +172,7 @@ the dependencies manually using anaconda followed by pulling hdbscan from pip:
 .. code:: bash
 
     conda install cython
+    conda install numpy scipy
     conda install scikit-learn
     pip install hdbscan
 
@@ -188,13 +195,28 @@ or
 
 .. code:: bash
 
-    conda install scikit-learn cython
+    conda install scikit-learn cython 
 
 Install the package
 
 .. code:: bash
 
     python setup.py install
+    
+----------------
+Help and Support
+----------------
+
+For simple issues you can consult the `FAQ <https://hdbscan.readthedocs.io/en/latest/faq.html>`_ in the documentation.
+If your issue is not suitably resolved there, please check the `issues <https://github.com/scikit-learn-contrib/hdbscan/issues>`_ on github. Finally, if no solution is available there feel free to `open an issue <https://github.com/scikit-learn-contrib/hdbscan/issues/new>`_ ; the authors will attempt to respond in a reasonably timely fashion.
+
+------------
+Contributing
+------------
+
+We welcome contributions in any form! Assistance with documentation, particularly expanding tutorials,
+is always welcome. To contribute please `fork the project <https://github.com/scikit-learn-contrib/hdbscan/issues#fork-destination-box>`_ make your changes and submit a pull request. We will do our best to work through any issues with
+you and get your code merged into the main branch.
 
 ---------
 Licensing
