### 2017-03-18 d11ed27
Add DOIs where they exist
