+++ paper/paper.bib
@@ -4,7 +4,9 @@
   booktitle={Pacific-Asia Conference on Knowledge Discovery and Data Mining},
   pages={160--172},
   year={2013},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-642-37456-2_14},
+  url={http://dx.doi.org/10.1007/978-3-642-37456-2_14}
 }
 
 @article{campello2015hierarchical,
@@ -15,7 +17,9 @@
   number={1},
   pages={5},
   year={2015},
-  publisher={ACM}
+  publisher={ACM},
+  url = {http://doi.acm.org/10.1145/2733381},
+  doi = {10.1145/2733381}
 }
 
 @article{chaudhuri2014consistent,
@@ -26,7 +30,8 @@
   number={12},
   pages={7900--7912},
   year={2014},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TIT.2014.2361055}
 }
 
 @inproceedings{chaudhuri2010rates,
@@ -38,7 +43,7 @@
  location = {Vancouver, British Columbia, Canada},
  pages = {343--351},
  numpages = {9},
- url = {http://dl.acm.org/citation.cfm?id=2997189.2997228},
+ url = {https://papers.nips.cc/paper/4068-rates-of-convergence-for-the-cluster-tree},
  acmid = {2997228},
  publisher = {Curran Associates Inc.},
  address = {USA},
