### 2017-03-18 #205 @lmcinnes
Hi @zhaozhang

Thanks for taking the time to look through everything. I've attempted to remedy each of these points. My only issue is that there does not appear to be a DOI for "Rates of convergence for the cluster tree" from NIPS 2010 (NIPS proceedings seem to forego DOIs for reasons that are not clear to me). Any recommendations on that front would be appreciated.
