### 2017-03-18 #205 @danielskatz
If there's no DOI for a reference, then you don't need to supply one :)

JOSS just wants to be sure that where there is a DOI, it's used.
