### 2017-03-20 #205 @zhaozhang
Hi @danielskatz

Yes, I have update the checkboxes. I think it is ready to be accepted.
