### 2017-03-20 #205 @danielskatz
@arfon, over to you for the next steps.

@zhaozhang, thanks for the quick work in suggesting changes

@lmcinnes, thanks for your quick work in responding to the suggestions.
