### 2017-03-20 #205 @lmcinnes
Thanks @zhaozhang and @danielskatz for steering me through this so well.

On Mon, Mar 20, 2017 at 5:08 PM, Daniel S. Katz <notifications@github.com>
wrote:

> @arfon <https://github.com/arfon>, over to you for the next steps.
>
> @zhaozhang <https://github.com/zhaozhang>, thanks for the quick work in
> suggesting changes
>
> @lmcinnes <https://github.com/lmcinnes>, thanks for your quick work in
> responding to the suggestions.
>
