### 2017-03-21 #205 @lmcinnes
I believe I managed to make this work: http://doi.org/10.5281/zenodo.401403

Thanks for all the help.

On Mon, Mar 20, 2017 at 8:04 PM, Arfon Smith <notifications@github.com>
wrote:

> @lmcinnes <https://github.com/lmcinnes> - At this point could you make an
> archive of the reviewed software in Zenodo/figshare/other service and
> update this thread with the DOI of the archive? I can then move forward
> with accepting the submission.
>
